+++ client/src/app/guides/page.tsx
@@ -6,6 +6,15 @@ export interface Guide {
   restaurantIds: number[];
 }
 
+const titlePrefix =
+  process.env.NODE_ENV === "development" ? "[Development] " : "";
+
+export const metadata = {
+  title: `${titlePrefix} Guides | Good Feeds`,
+  description:
+    "Recommended restaurants in Melbourne, Victoria, Australia and surrounding suburbs.",
+};
+
 async function getGuides(): Promise<{ [slug: string]: Guide }> {
   const response = await fetch(`${process.env.NEXT_CLIENT_URL}/guides.json`, {
     next: { revalidate: 10 },