+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add guides - Curated and sometimes themed recommendations for restaurants
+- Set specific page title on About and Guide pages
 
 ## [1.11.5] - 2025-11-14